requirements.txt
numpy
torch
torchvision
opencv-python
wandb
numexpr
scipy
einops
scikit-learn
pandas
matplotlib
Pillow
tqdm
ujson
timm==0.4.5
colour
bitsandbytes
torch_geometric
tqdm
